Add an AEnKF (Asynchronous Ensemble Kalman Filter) example.
This example extends the EnKF example. It takes the time series of four gauges as observations and modifies the lower zone storage of four subbasins. Hence, this is the first good test to show that our new spatial noise model works well with HydPy's updated ExchangeItem mechanisms. However, maybe we should consider adding more tests, for example, one working on the "sub-device" level.